requirements.txt
# AI Knowledge Workspace - Python 3.11 Compatible
# Fixed version conflicts

# Core RAG
langchain==0.1.20
langchain-community==0.0.38
langchain-openai==0.1.7

# PDF Processing
PyPDF2==3.0.1

# Vector Database
chromadb==0.4.24

# Embeddings (Fixed versions)
sentence-transformers==2.7.0
scikit-learn==1.3.0

# Data Processing (Compatible versions)
numpy==1.24.3
pandas==2.0.3

# Groq
groq==0.4.2

# Environment
python-dotenv==1.0.1

# API Framework
fastapi==0.111.0
uvicorn==0.29.0

# Utilities
tqdm==4.66.4
requests==2.31.0